Add FastAPI route for tools/fetch
