+++ lib/galaxy/webapps/galaxy/api/tools.py
@@ -1,15 +1,16 @@
 import logging
 import os
-from json import (
-    dumps,
-    loads,
-)
+from json import loads
 from typing import (
     Any,
     cast,
     Dict,
     Optional,
 )
+
+from fastapi import Request
+from fastapi.exceptions import RequestValidationError
+from pydantic import ValidationError
 
 from galaxy import (
     exceptions,
@@ -18,10 +19,9 @@ from galaxy import (
 )
 from galaxy.datatypes.data import get_params_and_input_name
 from galaxy.managers.collections import DatasetCollectionManager
-from galaxy.managers.collections_util import dictify_dataset_collection_instance
+from galaxy.managers.context import ProvidesHistoryContext
 from galaxy.managers.hdas import HDAManager
 from galaxy.managers.histories import HistoryManager
-from galaxy.model import PostJobAction
 from galaxy.schema.fetch_data import FetchDataPayload
 from galaxy.tools.evaluation import global_tool_errors
 from galaxy.util.zipstream import ZipstreamWrapper
@@ -34,12 +34,13 @@ from galaxy.web import (
 from galaxy.web.framework.decorators import expose_api_raw
 from galaxy.webapps.base.controller import UsesVisualizationMixin
 from galaxy.webapps.base.webapp import GalaxyWebTransaction
+from galaxy.webapps.galaxy.services.tools import ToolsService
 from . import (
     BaseGalaxyAPIController,
     depends,
+    DependsOnTrans,
     Router,
 )
-from ._fetch_util import validate_and_normalize_targets
 
 log = logging.getLogger(__name__)
 
@@ -53,6 +54,36 @@ SEARCH_RESERVED_TERMS_FAVORITES = ["#favs", "#favorites", "#favourites"]
 router = Router(tags=["tools"])
 
 
+@router.cbv
+class FastAPITools:
+    service: ToolsService = depends(ToolsService)
+
+    @router.post(
+        "/api/tools/fetch",
+        responses={
+            200: {
+                "content": {
+                    "application/json": {"schema": FetchDataPayload.schema()},
+                    "multipart/form-data": {"schema": FetchDataPayload.schema()},
+                }
+            }
+        },
+        summary="Upload files to Galaxy",
+    )
+    async def fetch(self, request: Request, trans: ProvidesHistoryContext = DependsOnTrans):
+        content_type = request.headers["content-type"]
+        if content_type == "application/json":
+            data = await request.json()
+        elif content_type.startswith("multipart/form-data"):
+            data = await request.form()
+        else:
+            raise exceptions.MessageException(f"Invalid content-type {content_type} for request")
+        try:
+            return self.service._create_fetch(trans, FetchDataPayload(**data))
+        except ValidationError as e:
+            raise RequestValidationError(e.raw_errors)
+
+
 class ToolsController(BaseGalaxyAPIController, UsesVisualizationMixin):
     """
     RESTful controller for interactions with tools.
@@ -61,6 +92,7 @@ class ToolsController(BaseGalaxyAPIController, UsesVisualizationMixin):
     history_manager: HistoryManager = depends(HistoryManager)
     hda_manager: HDAManager = depends(HDAManager)
     hdca_manager: DatasetCollectionManager = depends(DatasetCollectionManager)
+    service: ToolsService = depends(ToolsService)
 
     @expose_api_anonymous_and_sessionless
     def index(self, trans: GalaxyWebTransaction, **kwds):
@@ -96,12 +128,12 @@ class ToolsController(BaseGalaxyAPIController, UsesVisualizationMixin):
                 else:
                     hits = None
             else:
-                hits = self._search(q, view)
+                hits = self.service._search(q, view)
             results = []
             if hits:
                 for hit in hits:
                     try:
-                        tool = self._get_tool(hit, user=trans.user)
+                        tool = self.service._get_tool(trans, hit, user=trans.user)
                         if tool:
                             results.append(tool.id)
                     except exceptions.AuthenticationFailed:
@@ -112,7 +144,7 @@ class ToolsController(BaseGalaxyAPIController, UsesVisualizationMixin):
 
         # Find whether to detect.
         if tool_id:
-            detected_versions = self._detect(trans, tool_id)
+            detected_versions = self.service._detect(trans, tool_id)
             return detected_versions
 
         # Return everything.
@@ -141,7 +173,7 @@ class ToolsController(BaseGalaxyAPIController, UsesVisualizationMixin):
         io_details = util.string_as_bool(kwd.get("io_details", False))
         link_details = util.string_as_bool(kwd.get("link_details", False))
         tool_version = kwd.get("tool_version")
-        tool = self._get_tool(id, user=trans.user, tool_version=tool_version)
+        tool = self.service._get_tool(trans, id, user=trans.user, tool_version=tool_version)
         return tool.to_dict(trans, io_details=io_details, link_details=link_details)
 
     @expose_api_anonymous
@@ -158,7 +190,7 @@ class ToolsController(BaseGalaxyAPIController, UsesVisualizationMixin):
             history = self.history_manager.get_owned(
                 self.decode_id(history_id), trans.user, current_history=trans.history
             )
-        tool = self._get_tool(id, tool_version=tool_version, user=trans.user)
+        tool = self.service._get_tool(trans, id, tool_version=tool_version, user=trans.user)
         return tool.to_json(trans, kwd.get("inputs", kwd), history=history)
 
     @web.require_admin
@@ -169,7 +201,7 @@ class ToolsController(BaseGalaxyAPIController, UsesVisualizationMixin):
         """
         kwd = _kwd_or_payload(kwd)
         tool_version = kwd.get("tool_version", None)
-        tool = self._get_tool(id, tool_version=tool_version, user=trans.user)
+        tool = self.service._get_tool(trans, id, tool_version=tool_version, user=trans.user)
         path = tool.test_data_path(kwd.get("filename"))
         if path:
             return path
@@ -182,7 +214,7 @@ class ToolsController(BaseGalaxyAPIController, UsesVisualizationMixin):
         GET /api/tools/{tool_id}/test_data_download?tool_version={tool_version}&filename={filename}
         """
         tool_version = kwd.get("tool_version", None)
-        tool = self._get_tool(id, tool_version=tool_version, user=trans.user)
+        tool = self.service._get_tool(trans, id, tool_version=tool_version, user=trans.user)
         filename = kwd.get("filename")
         if filename is None:
             raise exceptions.ObjectNotFound("Test data filename not specified.")
@@ -250,7 +282,7 @@ class ToolsController(BaseGalaxyAPIController, UsesVisualizationMixin):
                 if not tool.allow_user_access(trans.user):
                     raise exceptions.AuthenticationFailed(f"Access denied, please login for tool with id '{id}'.")
         else:
-            tools = [self._get_tool(id, tool_version=tool_version, user=trans.user)]
+            tools = [self.service._get_tool(trans, id, tool_version=tool_version, user=trans.user)]
 
         test_defs = []
         for tool in tools:
@@ -288,7 +320,7 @@ class ToolsController(BaseGalaxyAPIController, UsesVisualizationMixin):
         Return the resolver status for a specific tool id.
         [{"status": "installed", "name": "hisat2", "versionless": false, "resolver_type": "conda", "version": "2.0.3", "type": "package"}]
         """
-        tool = self._get_tool(id, user=trans.user)
+        tool = self.service._get_tool(trans, id, user=trans.user)
         return tool.tool_requirements_status
 
     @web.require_admin
@@ -311,7 +343,7 @@ class ToolsController(BaseGalaxyAPIController, UsesVisualizationMixin):
             build_dependency_cache:  If true, attempts to cache dependencies for this tool
             force_rebuild:           If true and cache dir exists, attempts to delete cache dir
         """
-        tool = self._get_tool(id, user=trans.user)
+        tool = self.service._get_tool(trans, id, user=trans.user)
         tool._view.install_dependencies(tool.requirements, **kwds)
         if kwds.get("build_dependency_cache"):
             tool.build_dependency_cache(**kwds)
@@ -336,7 +368,7 @@ class ToolsController(BaseGalaxyAPIController, UsesVisualizationMixin):
 
             resolver_type: Use the dependency resolver of this resolver_type to install dependency
         """
-        tool = self._get_tool(id, user=trans.user)
+        tool = self.service._get_tool(trans, id, user=trans.user)
         tool._view.uninstall_dependencies(requirements=tool.requirements, **kwds)
         # TODO: rework resolver install system to log and report what has been done.
         return tool.tool_requirements_status
@@ -351,7 +383,7 @@ class ToolsController(BaseGalaxyAPIController, UsesVisualizationMixin):
         parameters:
             force_rebuild:           If true and chache dir exists, attempts to delete cache dir
         """
-        tool = self._get_tool(id)
+        tool = self.service._get_tool(trans, id)
         tool.build_dependency_cache(**kwds)
         # TODO: Should also have a more meaningful return.
         return tool.tool_requirements_status
@@ -368,7 +400,7 @@ class ToolsController(BaseGalaxyAPIController, UsesVisualizationMixin):
         def to_dict(x):
             return x.to_dict()
 
-        tool = self._get_tool(id, user=trans.user)
+        tool = self.service._get_tool(trans, id, user=trans.user)
         if hasattr(tool, "lineage"):
             lineage_dict = tool.lineage.to_dict()
         else:
@@ -392,68 +424,9 @@ class ToolsController(BaseGalaxyAPIController, UsesVisualizationMixin):
             "guid": tool.guid,
         }
 
-    def _detect(self, trans: GalaxyWebTransaction, tool_id):
-        """
-        Detect whether the tool with the given id is installed.
-
-        :param tool_id: exact id of the tool
-        :type tool_id:  str
-
-        :return:      list with available versions
-        "return type: list
-        """
-        tools = self.app.toolbox.get_tool(tool_id, get_all_versions=True)
-        detected_versions = []
-        if tools:
-            for tool in tools:
-                if tool and tool.allow_user_access(trans.user):
-                    detected_versions.append(tool.version)
-        return detected_versions
-
-    def _search(self, q, view):
-        """
-        Perform the search on the given query.
-        Boosts and numer of results are configurable in galaxy.ini file.
-
-        :param q: the query to search with
-        :type  q: str
-
-        :return:      Dictionary containing the tools' ids of the best hits.
-        :return type: dict
-        """
-        panel_view = view or self.app.config.default_panel_view
-        tool_name_boost = self.app.config.get("tool_name_boost", 9)
-        tool_id_boost = self.app.config.get("tool_id_boost", 9)
-        tool_section_boost = self.app.config.get("tool_section_boost", 3)
-        tool_description_boost = self.app.config.get("tool_description_boost", 2)
-        tool_label_boost = self.app.config.get("tool_label_boost", 1)
-        tool_stub_boost = self.app.config.get("tool_stub_boost", 5)
-        tool_help_boost = self.app.config.get("tool_help_boost", 0.5)
-        tool_search_limit = self.app.config.get("tool_search_limit", 20)
-        tool_enable_ngram_search = self.app.config.get("tool_enable_ngram_search", False)
-        tool_ngram_minsize = self.app.config.get("tool_ngram_minsize", 3)
-        tool_ngram_maxsize = self.app.config.get("tool_ngram_maxsize", 4)
-
-        results = self.app.toolbox_search.search(
-            q=q,
-            panel_view=panel_view,
-            tool_name_boost=tool_name_boost,
-            tool_id_boost=tool_id_boost,
-            tool_section_boost=tool_section_boost,
-            tool_description_boost=tool_description_boost,
-            tool_label_boost=tool_label_boost,
-            tool_stub_boost=tool_stub_boost,
-            tool_help_boost=tool_help_boost,
-            tool_search_limit=tool_search_limit,
-            tool_enable_ngram_search=tool_enable_ngram_search,
-            tool_ngram_minsize=tool_ngram_minsize,
-            tool_ngram_maxsize=tool_ngram_maxsize,
-        )
-        return results
-
     @expose_api_anonymous_and_sessionless
     def citations(self, trans: GalaxyWebTransaction, id, **kwds):
-        tool = self._get_tool(id, user=trans.user)
+        tool = self.service._get_tool(trans, id, user=trans.user)
         rval = []
         for citation in tool.citations:
             rval.append(citation.to_dict("bibtex"))
@@ -461,7 +434,7 @@ class ToolsController(BaseGalaxyAPIController, UsesVisualizationMixin):
 
     @expose_api
     def conversion(self, trans: GalaxyWebTransaction, tool_id, payload, **kwd):
-        converter = self._get_tool(tool_id, user=trans.user)
+        converter = self.service._get_tool(trans, tool_id, user=trans.user)
         target_type = payload.get("target_type")
         source_type = payload.get("source_type")
         input_src = payload.get("src")
@@ -504,11 +477,11 @@ class ToolsController(BaseGalaxyAPIController, UsesVisualizationMixin):
         # Make the target datatype available to the converter
         params["__target_datatype__"] = target_type
         vars = converter.handle_input(trans, params, history=target_history)
-        return self._handle_inputs_output_to_api_response(trans, converter, target_history, vars)
+        return self.service._handle_inputs_output_to_api_response(trans, converter, target_history, vars)
 
     @expose_api_anonymous_and_sessionless
     def xrefs(self, trans: GalaxyWebTransaction, id, **kwds):
-        tool = self._get_tool(id, user=trans.user)
+        tool = self.service._get_tool(trans, id, user=trans.user)
         return tool.xrefs
 
     @web.require_admin
@@ -527,42 +500,9 @@ class ToolsController(BaseGalaxyAPIController, UsesVisualizationMixin):
             raise exceptions.InsufficientPermissionsException(
                 "Only administrators may display tool sources on this Galaxy server."
             )
-        tool = self._get_tool(id, user=trans.user, tool_version=kwds.get("tool_version"))
+        tool = self.service._get_tool(trans, id, user=trans.user, tool_version=kwds.get("tool_version"))
         trans.response.headers["language"] = tool.tool_source.language
         return tool.tool_source.to_string()
-
-    # @router.post("/api/fetch", summary="Post content to upload into History Dataset")
-    @expose_api
-    def fetch(self, trans: GalaxyWebTransaction, payload: FetchDataPayload, **kwd):
-        """Adapt clean API to tool-constrained API."""
-        payload = FetchDataPayload(**payload).dict()
-        request_version = "1"
-        if "history_id" not in payload:
-            raise exceptions.RequestParameterMissingException("history_id must be specified")
-        history_id = payload.pop("history_id")
-        clean_payload = {}
-        files_payload = {}
-        for key, value in payload.items():
-            if key == "key":
-                continue
-            if key.startswith("files_") or key.startswith("__files_"):
-                files_payload[key] = value
-                continue
-            clean_payload[key] = value
-        validate_and_normalize_targets(trans, clean_payload)
-        clean_payload["check_content"] = trans.app.config.check_upload_content
-        request = dumps(clean_payload)
-        create_payload = {
-            "tool_id": "__DATA_FETCH__",
-            "history_id": history_id,
-            "inputs": {
-                "request_version": request_version,
-                "request_json": request,
-                "file_count": str(len(files_payload)),
-            },
-        }
-        create_payload.update(files_payload)
-        return self._create(trans, create_payload, **kwd)
 
     @web.require_admin
     @expose_api
@@ -594,179 +534,7 @@ class ToolsController(BaseGalaxyAPIController, UsesVisualizationMixin):
             )
         if tool_id is None and tool_uuid is None:
             raise exceptions.RequestParameterInvalidException("Must specify a valid tool_id to use this endpoint.")
-        return self._create(trans, payload, **kwd)
-
-    def _create(self, trans: GalaxyWebTransaction, payload, **kwd):
-        if trans.user_is_bootstrap_admin:
-            raise exceptions.RealUserRequiredException("Only real users can execute tools or run jobs.")
-        action = payload.get("action")
-        if action == "rerun":
-            raise Exception("'rerun' action has been deprecated")
-
-        # Get tool.
-        tool_version = payload.get("tool_version")
-        tool_id = payload.get("tool_id")
-        tool_uuid = payload.get("tool_uuid")
-        get_kwds = dict(
-            tool_id=tool_id,
-            tool_uuid=tool_uuid,
-            tool_version=tool_version,
-        )
-        if tool_id is None and tool_uuid is None:
-            raise exceptions.RequestParameterMissingException("Must specify either a tool_id or a tool_uuid.")
-
-        tool = trans.app.toolbox.get_tool(**get_kwds)
-        if not tool:
-            log.debug(f"Not found tool with kwds [{get_kwds}]")
-            raise exceptions.ToolMissingException("Tool not found.")
-        if not tool.allow_user_access(trans.user):
-            raise exceptions.ItemAccessibilityException("Tool not accessible.")
-        if trans.app.config.user_activation_on:
-            if not trans.user:
-                log.warning("Anonymous user attempts to execute tool, but account activation is turned on.")
-            elif not trans.user.active:
-                log.warning(
-                    f'User "{trans.user.email}" attempts to execute tool, but account activation is turned on and user account is not active.'
-                )
-
-        # Set running history from payload parameters.
-        # History not set correctly as part of this API call for
-        # dataset upload.
-        history_id = payload.get("history_id")
-        if history_id:
-            decoded_id = self.decode_id(history_id)
-            target_history = self.history_manager.get_owned(decoded_id, trans.user, current_history=trans.history)
-        else:
-            target_history = None
-
-        # Set up inputs.
-        inputs = payload.get("inputs", {})
-        if not isinstance(inputs, dict):
-            raise exceptions.RequestParameterInvalidException(f"inputs invalid {inputs}")
-
-        # Find files coming in as multipart file data and add to inputs.
-        for k, v in payload.items():
-            if k.startswith("files_") or k.startswith("__files_"):
-                inputs[k] = v
-
-        # for inputs that are coming from the Library, copy them into the history
-        self._patch_library_inputs(trans, inputs, target_history)
-
-        # TODO: encode data ids and decode ids.
-        # TODO: handle dbkeys
-        params = util.Params(inputs, sanitize=False)
-        incoming = params.__dict__
-
-        # use_cached_job can be passed in via the top-level payload or among the tool inputs.
-        # I think it should be a top-level parameter, but because the selector is implemented
-        # as a regular tool parameter we accept both.
-        use_cached_job = payload.get("use_cached_job", False) or util.string_as_bool(
-            inputs.get("use_cached_job", "false")
-        )
-
-        input_format = str(payload.get("input_format", "legacy"))
-
-        vars = tool.handle_input(
-            trans, incoming, history=target_history, use_cached_job=use_cached_job, input_format=input_format
-        )
-
-        new_pja_flush = False
-        for job in vars.get("jobs", []):
-            if inputs.get("send_email_notification", False):
-                # Unless an anonymous user is invoking this via the API it
-                # should never be an option, but check and enforce that here
-                if trans.user is None:
-                    raise exceptions.ToolExecutionError("Anonymously run jobs cannot send an email notification.")
-                else:
-                    job_email_action = PostJobAction("EmailAction")
-                    job.add_post_job_action(job_email_action)
-                    new_pja_flush = True
-
-        if new_pja_flush:
-            trans.sa_session.flush()
-
-        return self._handle_inputs_output_to_api_response(trans, tool, target_history, vars)
-
-    def _handle_inputs_output_to_api_response(self, trans, tool, target_history, vars):
-        # TODO: check for errors and ensure that output dataset(s) are available.
-        output_datasets = vars.get("out_data", [])
-        rval: Dict[str, Any] = {"outputs": [], "output_collections": [], "jobs": [], "implicit_collections": []}
-        rval["produces_entry_points"] = tool.produces_entry_points
-        job_errors = vars.get("job_errors", [])
-        if job_errors:
-            # If we are here - some jobs were successfully executed but some failed.
-            rval["errors"] = job_errors
-
-        outputs = rval["outputs"]
-        # TODO:?? poss. only return ids?
-        for output_name, output in output_datasets:
-            output_dict = output.to_dict()
-            # add the output name back into the output data structure
-            # so it's possible to figure out which newly created elements
-            # correspond with which tool file outputs
-            output_dict["output_name"] = output_name
-            outputs.append(trans.security.encode_dict_ids(output_dict, skip_startswith="metadata_"))
-
-        for job in vars.get("jobs", []):
-            rval["jobs"].append(self.encode_all_ids(trans, job.to_dict(view="collection"), recursive=True))
-
-        for output_name, collection_instance in vars.get("output_collections", []):
-            history = target_history or trans.history
-            output_dict = dictify_dataset_collection_instance(
-                collection_instance,
-                security=trans.security,
-                url_builder=trans.url_builder,
-                parent=history,
-            )
-            output_dict["output_name"] = output_name
-            rval["output_collections"].append(output_dict)
-
-        for output_name, collection_instance in vars.get("implicit_collections", {}).items():
-            history = target_history or trans.history
-            output_dict = dictify_dataset_collection_instance(
-                collection_instance,
-                security=trans.security,
-                url_builder=trans.url_builder,
-                parent=history,
-            )
-            output_dict["output_name"] = output_name
-            rval["implicit_collections"].append(output_dict)
-
-        return rval
-
-    def _patch_library_inputs(self, trans: GalaxyWebTransaction, inputs, target_history):
-        """
-        Transform inputs from the data library to history items.
-        """
-        for k, v in inputs.items():
-            new_value = self._patch_library_dataset(trans, v, target_history)
-            if new_value:
-                v = new_value
-            elif isinstance(v, dict) and "values" in v:
-                for index, value in enumerate(v["values"]):
-                    patched = self._patch_library_dataset(trans, value, target_history)
-                    if patched:
-                        v["values"][index] = patched
-            inputs[k] = v
-
-    def _patch_library_dataset(self, trans: GalaxyWebTransaction, v, target_history):
-        if isinstance(v, dict) and "id" in v and v.get("src") == "ldda":
-            ldda = trans.sa_session.query(trans.app.model.LibraryDatasetDatasetAssociation).get(self.decode_id(v["id"]))
-            if trans.user_is_admin or trans.app.security_agent.can_access_dataset(
-                trans.get_current_user_roles(), ldda.dataset
-            ):
-                return ldda.to_history_dataset_association(target_history, add_to_history=True)
-
-    #
-    # -- Helper methods --
-    #
-    def _get_tool(self, id, tool_version=None, user=None):
-        tool = self.app.toolbox.get_tool(id, tool_version)
-        if not tool:
-            raise exceptions.ObjectNotFound(f"Could not find tool with id '{id}'.")
-        if not tool.allow_user_access(user):
-            raise exceptions.AuthenticationFailed(f"Access denied, please login for tool with id '{id}'.")
-        return tool
+        return self.service._create(trans, payload, **kwd)
 
 
 def _kwd_or_payload(kwd: Dict[str, Any]) -> Dict[str, Any]:
