+++ lib/galaxy/webapps/galaxy/services/_fetch_util.py
