+++ test/integration/test_repository_operations.py
@@ -111,17 +111,23 @@ class TestRepositoryInstallIntegrationTestCase(integration_util.IntegrationTestC
         assert int(response["ctx_rev"]) >= 4
 
     def _uninstall_repository(self):
-        tool = self._get("/api/tools/collection_column_join").json()
+        tool = self.get_tool()
         assert tool["version"] == "0.0.2"
         self.uninstall_repository(REPO.owner, REPO.name, REPO.changeset)
-        response = self._get("/api/tools/collection_column_join").json()
+        response = self.get_tool(assert_ok=False)
         assert response["err_msg"]
 
     def _install_repository(self, revision=None, version="0.0.2", verify_tool_absent=True):
         if verify_tool_absent:
-            response = self._get("/api/tools/collection_column_join").json()
+            response = self.get_tool(assert_ok=False)
             assert response["err_msg"]
         install_response = self.install_repository(REPO.owner, REPO.name, revision or REPO.changeset)
-        tool = self._get("/api/tools/collection_column_join").json()
-        assert tool["version"] == version
+        tool = self.get_tool()
+        assert tool.get("version") == version, tool
         return install_response
+
+    def get_tool(self, assert_ok=True):
+        response = self._get("/api/tools/collection_column_join")
+        if assert_ok:
+            self._assert_status_code_is_ok(response)
+        return response.json()
